+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Removed
 
+- `init_config` from all the tests.
 - `fetch_from_test` argument
 
 ## Fixed